refactor(core): remove dead exports and relocate test-only machinery
Generated-by: Codex
