+++ packages/core/src/__tests__/additional-permissions.test.ts
@@ -1,5 +1,5 @@
 import { describe, test } from 'node:test';
-import { expect } from '../test-helpers.js';
+import { expect } from './test-helpers.js';
 import {
   compactAdditionalFileSystemPermissions,
   serializeAdditionalPermissionProfile,
