+++ packages/core/src/__tests__/permission-profile.test.ts
@@ -1,5 +1,5 @@
 import { describe, test } from 'node:test';
-import { expect } from '../test-helpers.js';
+import { expect } from './test-helpers.js';
 import {
   canReadPath,
   canWritePath,
