+++ packages/core/src/settings/__tests__/sensitive.test.ts
@@ -1,5 +1,5 @@
 import { describe, test } from 'node:test';
-import { expect } from '../../test-helpers.js';
+import { expect } from '../../__tests__/test-helpers.js';
 import { applySensitivePatch, maskSensitive } from '../network-settings.js';
 
 describe('applySensitivePatch', () => {
