--- packages/core/src/settings/result.ts
@@ -1,1 +1,0 @@
-export { err, ok, tryResult, type Result } from '../result.js';
